feat: complete epic-04 operation synthesis decomposition
Split monolithic operation.py (871 lines) into 6 modules:
orchestrator.py (209), pipeline.py (284), resolution.py (213),
cache.py (61), export.py (136). All 82 tests pass.
