+++ plans/dessem-overhaul/.implementation-state.json
@@ -1,6 +1,6 @@
 {
   "plan": "dessem-overhaul",
-  "last_updated": "2026-03-16T23:25:04Z",
+  "last_updated": "2026-03-16T23:48:40Z",
   "progressive": true,
   "tickets": {
     "ticket-001-update-pyproject-and-python-version.md": {
@@ -63,7 +63,7 @@
           "dependency_clarity": 1.0,
           "atomicity": 1.0
         },
-        "quality": 0.90,
+        "quality": 0.9,
         "quality_breakdown": {
           "conformance": 1.0,
           "scope_adherence": 1.0,
@@ -159,7 +159,7 @@
           "dependency_clarity": 1.0,
           "atomicity": 1.0
         },
-        "quality": 0.90,
+        "quality": 0.9,
         "quality_breakdown": {
           "conformance": 1.0,
           "scope_adherence": 1.0,
@@ -340,20 +340,100 @@
       "workflow_step": null
     },
     "ticket-015-create-operation-synthesis-package.md": {
-      "status": "pending",
-      "detail_level": "outline"
+      "status": "completed",
+      "detail_level": "refined",
+      "scores": {
+        "readiness": 1.0,
+        "readiness_breakdown": {
+          "structure": 1.0,
+          "testability": 1.0,
+          "boundary": 1.0,
+          "dependency_clarity": 1.0,
+          "atomicity": 1.0
+        },
+        "quality": 0.95,
+        "quality_breakdown": {
+          "conformance": 1.0,
+          "scope_adherence": 1.0,
+          "lint_cleanliness": 1.0,
+          "type_safety": 0.75,
+          "test_delta": 1.0
+        }
+      },
+      "workflow_step": null,
+      "agent": "direct"
     },
     "ticket-016-extract-operation-cache-and-export.md": {
-      "status": "pending",
-      "detail_level": "outline"
+      "status": "completed",
+      "detail_level": "refined",
+      "scores": {
+        "readiness": 1.0,
+        "readiness_breakdown": {
+          "structure": 1.0,
+          "testability": 1.0,
+          "boundary": 1.0,
+          "dependency_clarity": 1.0,
+          "atomicity": 1.0
+        },
+        "quality": 0.95,
+        "quality_breakdown": {
+          "conformance": 1.0,
+          "scope_adherence": 1.0,
+          "lint_cleanliness": 1.0,
+          "type_safety": 0.75,
+          "test_delta": 1.0
+        }
+      },
+      "agent": "python-task-automation-developer",
+      "workflow_step": null
     },
     "ticket-017-extract-resolution-modules.md": {
-      "status": "pending",
-      "detail_level": "outline"
+      "status": "completed",
+      "detail_level": "refined",
+      "scores": {
+        "readiness": 1.0,
+        "readiness_breakdown": {
+          "structure": 1.0,
+          "testability": 1.0,
+          "boundary": 1.0,
+          "dependency_clarity": 1.0,
+          "atomicity": 1.0
+        },
+        "quality": 0.95,
+        "quality_breakdown": {
+          "conformance": 1.0,
+          "scope_adherence": 1.0,
+          "lint_cleanliness": 1.0,
+          "type_safety": 0.75,
+          "test_delta": 1.0
+        }
+      },
+      "agent": "python-task-automation-developer",
+      "workflow_step": null
     },
     "ticket-018-create-operation-orchestrator.md": {
-      "status": "pending",
-      "detail_level": "outline"
+      "status": "completed",
+      "detail_level": "refined",
+      "scores": {
+        "readiness": 0.98,
+        "readiness_breakdown": {
+          "structure": 1.0,
+          "testability": 1.0,
+          "boundary": 1.0,
+          "dependency_clarity": 1.0,
+          "atomicity": 0.8
+        },
+        "quality": 0.95,
+        "quality_breakdown": {
+          "conformance": 1.0,
+          "scope_adherence": 1.0,
+          "lint_cleanliness": 1.0,
+          "type_safety": 0.75,
+          "test_delta": 1.0
+        }
+      },
+      "agent": "python-task-automation-developer",
+      "workflow_step": null
     },
     "ticket-019-add-polars-dependency.md": {
       "status": "pending",
@@ -406,11 +486,11 @@
       "learning_extracted": true
     },
     "epic-03-deck-decomposition": {
-      "phase": "executing",
-      "learning_extracted": false
+      "phase": "completed",
+      "learning_extracted": true
     },
     "epic-04-operation-decomposition": {
-      "phase": "outline",
+      "phase": "executing",
       "learning_extracted": false
     },
     "epic-05-polars-migration": {
